Updated About component
