+++ app/components/About/ ScoreExplanation.js
@@ -1,0 +1,9 @@
+export default function ScoreExplanation() {
+    return (
+      <p className="text-gray-700 leading-relaxed text-lg font-light pt-2">
+        <span className="font-bold text-indigo-600">
+          The higher the score, the faster the browser.
+        </span>
+      </p>
+    );
+}
