+++ app/components/About/TestProcedure.js
@@ -1,0 +1,19 @@
+export default function TestProcedure() {
+    return (
+      <p className="text-gray-700 leading-relaxed text-lg font-light pt-2">
+        For each browser,{" "}
+        <span className="font-medium text-indigo-600">
+          five tests were conducted
+        </span>
+        . The{" "}
+        <span className="font-medium text-indigo-600">
+          best and worst results were eliminated
+        </span>
+        , and the{" "}
+        <span className="font-medium text-indigo-600">
+          average of the remaining three tests was calculated
+        </span>{" "}
+        to determine the final result.
+      </p>
+    );
+  }
